firewall_governor/requirements.txt
# Firewall Governor — Python Dependencies
# Install with: pip install -r requirements.txt

# ── API Framework ──────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# ── Config & Data ──────────────────────────────────────────────────────────
pyyaml==6.0.1
numpy==1.26.4

# ── Hardware ───────────────────────────────────────────────────────────────
viam-sdk==0.21.0

# ── Temporal Logic (LTL/STL invariant checking) ────────────────────────────
# NOTE: Verify rtamt latest version on PyPI before installing.
# rtamt 0.1.0 does not exist. As of 2026, latest stable is ~0.3.x.
# Run: pip index versions rtamt  — to check what is available.
rtamt==0.3.3

# ── Audio Pipeline ─────────────────────────────────────────────────────────
# sounddevice: microphone capture
sounddevice==0.4.6
# librosa: REQUIRED for resampling mic audio (44.1kHz) to Whisper format (16kHz)
# Without this, Whisper transcription will be garbled or error.
librosa==0.10.1
# openai-whisper: Whisper model and log-Mel spectrogram utilities
# NOTE: The original requirements.txt listed 'whisper-tiny==0.1.0' which does NOT exist on PyPI.
# The correct package name is 'openai-whisper'.
openai-whisper==20231117

# ── Semantic Similarity (Audio Bridge — Stage 3 validation) ───────────────
# Used by SemanticAudioBridge.compute_semantic_similarity().
# Replaces RapidFuzz fuzzy matching (character-level → wrong for semantic alignment).
sentence-transformers==2.7.0

# ── Security (HITL token generation) ──────────────────────────────────────
# secrets is a Python stdlib module — no install needed.
# Listed here for documentation clarity only.
# secrets (stdlib)
